@@ -4,13 +4,15 @@
   "description": "My simple redux application.",
   "scripts": {
     "clean": "rimraf dist",
-    "build:webpack": "cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
+    "build:webpack": "cross-env NODE_ENV=production webpack -p --config webpack.config.prod.js",
     "build": "npm run clean && npm run build:webpack",
     "build:static": "npm run build && npm run assets",
     "assets": "copy index.html dist && copy static/* dist && copy favicon.ico dist",
     "dev": "NODE_ENV=development npm start",
     "start": "node devServer.js",
-    "lint": "eslint src"
+    "lint": "eslint src --ext .js --ext .jsx",
+    "stats": "cross-env NODE_ENV=production webpack --config webpack.config.prod.js --json | webpack-bundle-size-analyzer",
+    "start:prod": "npm run build:static && pushstate-server dist"
   },
   "repository": {
     "type": "git",
@@ -23,7 +25,7 @@
     "babel-core": "^6.3.15",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.8",
-    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-latest": "^6.16.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-react-hmre": "^1.0.0",
     "cross-env": "^3.1.3",
@@ -31,6 +33,8 @@
     "eslint": "^3.10.2",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-babel": "^4.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.7.1",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.13.3",
@@ -43,16 +47,20 @@
     "rimraf": "^2.4.3",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.0",
-    "webpack": "^1.13.3",
-    "webpack-dev-middleware": "^1.4.0",
+    "webpack": "^2.1.0-beta.25",
+    "webpack-dev-middleware": "^1.8.4",
     "webpack-hot-middleware": "^2.13.2"
   },
   "dependencies": {
     "copy": "^0.3.0",
+    "preact": "^6.4.0",
+    "preact-compat": "^3.9.2",
+    "rc-animate": "^2.3.1",
     "react": "^15.4.0",
     "react-addons-css-transition-group": "^15.3.2",
     "react-dom": "^15.4.0",
     "react-redux": "^4.4.6",
+    "redux": "^3.6.0",
     "redux-thunk": "^2.1.0"
   }
 }
