@@ -1,265 +1,265 @@
-export const LINKS_LIST = {
-	democrats: {
-		donate: [
-			{
-				text: "DEMOCRATIC NATIONAL COMMITTEE",
-				link: "https://my.democrats.org/page/contribute/donate-to-help-democrats-today"
-			},
-			{
-				text: "DEMOCRATIC SENATORIAL CAMPAIGN COMMITTEE",
-				link: "https://secure.dscc.org/page/contribute"
-			},
-			{
-				text: "DEMOCRATIC CONGRESSIONAL CAMPAIGN COMMITTEE",
-				link: "https://secure.actblue.com/contribute/page/dcccactblue"
-			},
-			{
-				text: "SHE SHOULD RUN",
-				link: "https://sheshouldrun.nationbuilder.com/donate"
-			},
-			{
-				text: "EMILY'S LIST",
-				link: "https://secure.emilyslist.org/page/contribute/donate-to-emily"
-			}
-		],
-		volunteer: [
-			{
-				text: "DEMOCRATIC NATIONAL COMMITTEE",
-				link: "http://my.democrats.org/page/s/help-elect-democrats"
-			},
-			{
-				text: "DEMOCRATIC CONGRESSIONAL CAMPAIGN COMMITTEE",
-				link: "http://action.dccc.org/page/s/volunteer"
-			},
-			{
-				text: "DEMOCRATIC LEGISLATIVE CAMPAIGN COMMITTEE",
-				link: "http://www.dlcc.org/volunteer"
-			},
-			{
-				text: "WHOEVER IS RUNNING FOR SHIT IN YOUR STATE",
-				link: "https://www.google.com/webhp?sourceid=chrome-instant&ion=1&espv=2&ie=UTF-8#q=who%20is%20running%20for%20office%20in%202018"
-			}
-		]
-	},
+export default {
+  democrats: {
+    donate: [
+      {
+        text: 'DEMOCRATIC NATIONAL COMMITTEE',
+        link: 'https://my.democrats.org/page/contribute/donate-to-help-democrats-today'
+      },
+      {
+        text: 'DEMOCRATIC SENATORIAL CAMPAIGN COMMITTEE',
+        link: 'https://secure.dscc.org/page/contribute'
+      },
+      {
+        text: 'DEMOCRATIC CONGRESSIONAL CAMPAIGN COMMITTEE',
+        link: 'https://secure.actblue.com/contribute/page/dcccactblue'
+      },
+      {
+        text: 'SHE SHOULD RUN',
+        link: 'https://sheshouldrun.nationbuilder.com/donate'
+      },
+      {
+        text: "EMILY'S LIST",
+        link: 'https://secure.emilyslist.org/page/contribute/donate-to-emily'
+      }
+    ],
+    volunteer: [
+      {
+        text: 'DEMOCRATIC NATIONAL COMMITTEE',
+        link: 'http://my.democrats.org/page/s/help-elect-democrats'
+      },
+      {
+        text: 'DEMOCRATIC CONGRESSIONAL CAMPAIGN COMMITTEE',
+        link: 'http://action.dccc.org/page/s/volunteer'
+      },
+      {
+        text: 'DEMOCRATIC LEGISLATIVE CAMPAIGN COMMITTEE',
+        link: 'http://www.dlcc.org/volunteer'
+      },
+      {
+        text: 'WHOEVER IS RUNNING FOR SHIT IN YOUR STATE',
+        link: 'https://www.google.com/webhp?sourceid=chrome-instant&ion=1&espv=2&ie=UTF-8#q=who%20is%20running%20for%20office%20in%202018'
+      }
+    ]
+  },
 
-	lgbtq: {
-		donate: [
-			{
-				text: "THE ACLU",
-				link: "https://action.aclu.org/donate-aclu?ms=web_horiz_nav_hp"
-			},
-			{
-				text: "THE HUMAN RIGHTS CAMPAIGN",
-				link: "https://give.hrc.org/ea-action/action?ea.client.id=1954&ea.campaign.id=51848"
-			},
-			{
-				text: "LAMBDA LEGAL",
-				link: "https://www.lambdalegal.org/donate"
-			},
-			{
-				text: "SOUTHERNERS ON NEW GROUND",
-				link: "https://song.ourpowerbase.net/civicrm/contribute/transact?reset=1&id=2"
-			},
-			{
-				text: "GLAAD",
-				link: "https://glaad.nationbuilder.com/donate"
-			},
-			{
-				text: "THE TREVOR PROJECT",
-				link: "https://give.thetrevorproject.org/checkout/donation?eid=63307"
-			},
-			{
-				text: "SILVIA RIVERA LAW PROJECT",
-				link: "http://srlp.org/srlp/?page=CiviCRM&q=civicrm/contribute/transact&reset=1&id=4"
-			},
-			{
-				text: "EVERYONE IS GAY",
-				link: "http://everyoneisgay.com/donate/"
-			}
-		],
-		volunteer: [
-			{
-				text: "GLAAD",
-				link: "http://www.glaad.org/getinvolved/volunteer"
-			},
-			{
-				text: "MAZZONI CENTER",
-				link: "https://www.mazzonicenter.org/donate-now/volunteer-opportunities"
-			},
-			{
-				text: "THE SYLVIA RIVERA LAW PROJECT",
-				link: "http://srlp.org/get-involved/volunteer/"
-			},
-			{
-				text: "THE TREVOR PROJECT",
-				link: "http://www.thetrevorproject.org/pages/volunteer"
-			},
-			{
-				text: "SOUTHERNERS ON NEW GROUND",
-				link: "https://song.ourpowerbase.net/civicrm/contribute/transact?reset=1&id=2"
-			}
+  lgbtq: {
+    donate: [
+      {
+        text: 'THE ACLU',
+        link: 'https://action.aclu.org/donate-aclu?ms=web_horiz_nav_hp'
+      },
+      {
+        text: 'THE HUMAN RIGHTS CAMPAIGN',
+        link: 'https://give.hrc.org/ea-action/action?ea.client.id=1954&ea.campaign.id=51848'
+      },
+      {
+        text: 'LAMBDA LEGAL',
+        link: 'https://www.lambdalegal.org/donate'
+      },
+      {
+        text: 'SOUTHERNERS ON NEW GROUND',
+        link: 'https://song.ourpowerbase.net/civicrm/contribute/transact?reset=1&id=2'
+      },
+      {
+        text: 'GLAAD',
+        link: 'https://glaad.nationbuilder.com/donate'
+      },
+      {
+        text: 'THE TREVOR PROJECT',
+        link: 'https://give.thetrevorproject.org/checkout/donation?eid=63307'
+      },
+      {
+        text: 'SILVIA RIVERA LAW PROJECT',
+        link: 'http://srlp.org/srlp/?page=CiviCRM&q=civicrm/contribute/transact&reset=1&id=4'
+      },
+      {
+        text: 'EVERYONE IS GAY',
+        link: 'http://everyoneisgay.com/donate/'
+      }
+    ],
+    volunteer: [
+      {
+        text: 'GLAAD',
+        link: 'http://www.glaad.org/getinvolved/volunteer'
+      },
+      {
+        text: 'MAZZONI CENTER',
+        link: 'https://www.mazzonicenter.org/donate-now/volunteer-opportunities'
+      },
+      {
+        text: 'THE SYLVIA RIVERA LAW PROJECT',
+        link: 'http://srlp.org/get-involved/volunteer/'
+      },
+      {
+        text: 'THE TREVOR PROJECT',
+        link: 'http://www.thetrevorproject.org/pages/volunteer'
+      },
+      {
+        text: 'SOUTHERNERS ON NEW GROUND',
+        link: 'https://song.ourpowerbase.net/civicrm/contribute/transact?reset=1&id=2'
+      }
 
-		]
-	},
+    ]
+  },
 
-	racism: {
-		donate: [
-			{
-				text: "BLACK LIVES MATTER",
-				link: "https://donate.idex.org/checkout/donation?eid=66399"
-			},
-			{
-				text: "COUNCIL ON AMERICAN-ISLAMIC RELATIONS",
-				link: "https://www.cair.com/donations/general-donation/campaign/#/donation"
-			},
-			{
-				text: "THE NAACP",
-				link: "https://donate.naacp.org/page/contribute/make-a-contribution"
-			},
-			{
-				text: "NATIONAL IMMIGRATION LAW CENTER",
-				link: "https://nilc.z2systems.com/np/clients/nilc/donation.jsp?campaign=15&"
-			},
-			{
-				text: "THE SOUTHERN POVERTY LAW CENTER",
-				link: "https://donate.splcenter.org/sslpage.aspx?pid=463"
-			},
-			{
-				text: "THE ANTI DEFAMATION LEAGUE",
-				link: "https://secure2.convio.net/adl/site/SPageNavigator/donation-api/donation_form_single_main_1.html"
-			},
-			{
-				text: "THE ACLU",
-				link: "https://action.aclu.org/donate-aclu?ms=web_horiz_nav_hp"
-			},
-			{
-				text: "SHOWING UP FOR RACIAL JUSTICE",
-				link: "https://showingupforracialjustice.nationbuilder.com/donate"
-			},
-			{
-				text: "NATIVE AMERICAN RIGHTS FUND",
-				link: "https://secure2.convio.net/narf/site/Donation2?idb=1549719056&DONATION_LEVEL_ID_SELECTED=1&df_id=1641&1641.donation=form1&mfc_pref=T&idb=0"
-			}
+  racism: {
+    donate: [
+      {
+        text: 'BLACK LIVES MATTER',
+        link: 'https://donate.idex.org/checkout/donation?eid=66399'
+      },
+      {
+        text: 'COUNCIL ON AMERICAN-ISLAMIC RELATIONS',
+        link: 'https://www.cair.com/donations/general-donation/campaign/#/donation'
+      },
+      {
+        text: 'THE NAACP',
+        link: 'https://donate.naacp.org/page/contribute/make-a-contribution'
+      },
+      {
+        text: 'NATIONAL IMMIGRATION LAW CENTER',
+        link: 'https://nilc.z2systems.com/np/clients/nilc/donation.jsp?campaign=15&'
+      },
+      {
+        text: 'THE SOUTHERN POVERTY LAW CENTER',
+        link: 'https://donate.splcenter.org/sslpage.aspx?pid=463'
+      },
+      {
+        text: 'THE ANTI DEFAMATION LEAGUE',
+        link: 'https://secure2.convio.net/adl/site/SPageNavigator/donation-api/donation_form_single_main_1.html'
+      },
+      {
+        text: 'THE ACLU',
+        link: 'https://action.aclu.org/donate-aclu?ms=web_horiz_nav_hp'
+      },
+      {
+        text: 'SHOWING UP FOR RACIAL JUSTICE',
+        link: 'https://showingupforracialjustice.nationbuilder.com/donate'
+      },
+      {
+        text: 'NATIVE AMERICAN RIGHTS FUND',
+        link: 'https://secure2.convio.net/narf/site/Donation2?idb=1549719056&DONATION_LEVEL_ID_SELECTED=1&df_id=1641&1641.donation=form1&mfc_pref=T&idb=0'
+      }
 
-		],
+    ],
 
-		volunteer: [
-			{
-				text: "CAMPAIGN ZERO",
-				link: "http://www.staywoke.org/"
-			},
-			{
-				text: "THE ANTI DEFAMATION LEAGUE",
-				link: "http://www.adl.org/regions/"
-			},
-			{
-				text: "THE NAACP",
-				link: "http://www.naacp.org/find-local-unit/"
-			},
-			{
-				text: "COUNCIL ON AMERICAN-ISLAMIC RELATIONS",
-				link: "https://www.cair.com/cair-chapters.html"
-			},
-			{
-				text: "SHOWING UP FOR RACIAL JUSTICE",
-				link: "http://www.showingupforracialjustice.org/organize"
-			}
-		]
-	},
+    volunteer: [
+      {
+        text: 'CAMPAIGN ZERO',
+        link: 'http://www.staywoke.org/'
+      },
+      {
+        text: 'THE ANTI DEFAMATION LEAGUE',
+        link: 'http://www.adl.org/regions/'
+      },
+      {
+        text: 'THE NAACP',
+        link: 'http://www.naacp.org/find-local-unit/'
+      },
+      {
+        text: 'COUNCIL ON AMERICAN-ISLAMIC RELATIONS',
+        link: 'https://www.cair.com/cair-chapters.html'
+      },
+      {
+        text: 'SHOWING UP FOR RACIAL JUSTICE',
+        link: 'http://www.showingupforracialjustice.org/organize'
+      }
+    ]
+  },
 
-	women: {
-		donate: [
-			{
-				text: "NOW",
-				link: "https://salsa.wiredforchange.com/o/5996/p/salsa/donation/common/public/?donate_page_KEY=11005"
-			},
-			{
-				text: "NARAL",
-				link: "https://actnow.prochoiceamerica.org/donate/supportnaral/?source=leftbar#.WCNXOOErKgQ"
-			},
-			{
-				text: "PLANNED PARENTHOOD",
-				link: "https://secure.ppaction.org/site/Donation2?df_id=12913&12913.donation=form1&_ga=1.38484843.997121566.1477408976"
-			},
-			{
-				text: "RAINN",
-				link: "https://donate.rainn.org/"
-			},
-			{
-				text: "NATIONAL COALITION AGAINST DOMESTIC VIOLENCE",
-				link: "http://ncadv.org/give-back/donate"
-			},
-			{
-				text: "REPRODUCTIVE HEALTH ACCESS PROJECT",
-				link: "https://donatenow.networkforgood.org/1438163"
-			},
-			{
-				text: "CENTER FOR REPRODUCTIVE RIGHTS",
-				link: "https://secure3.convio.net/cfrr/site/Donation2;jsessionid=97A4C7A9488B5DDF3450546DE1123DB2.app347a?1380.donation=form1&df_id=1380&web&_ga=1.145398384.513337705.1476995534"
-			}
-		],
-		volunteer: [
-			{
-				text: "NOW",
-				link: "http://nownyc.org/get-involved/"
-			},
-			{
-				text: "NARAL",
-				link: "http://actnow.prochoiceamerica.org/signup/volunteer/?source=website#.WCNXKuErKgQ"
-			},
+  women: {
+    donate: [
+      {
+        text: 'NOW',
+        link: 'https://salsa.wiredforchange.com/o/5996/p/salsa/donation/common/public/?donate_page_KEY=11005'
+      },
+      {
+        text: 'NARAL',
+        link: 'https://actnow.prochoiceamerica.org/donate/supportnaral/?source=leftbar#.WCNXOOErKgQ'
+      },
+      {
+        text: 'PLANNED PARENTHOOD',
+        link: 'https://secure.ppaction.org/site/Donation2?df_id=12913&12913.donation=form1&_ga=1.38484843.997121566.1477408976'
+      },
+      {
+        text: 'RAINN',
+        link: 'https://donate.rainn.org/'
+      },
+      {
+        text: 'NATIONAL COALITION AGAINST DOMESTIC VIOLENCE',
+        link: 'http://ncadv.org/give-back/donate'
+      },
+      {
+        text: 'REPRODUCTIVE HEALTH ACCESS PROJECT',
+        link: 'https://donatenow.networkforgood.org/1438163'
+      },
+      {
+        text: 'CENTER FOR REPRODUCTIVE RIGHTS',
+        link: 'https://secure3.convio.net/cfrr/site/Donation2;jsessionid=97A4C7A9488B5DDF3450546DE1123DB2.app347a?1380.donation=form1&df_id=1380&web&_ga=1.145398384.513337705.1476995534'
+      }
+    ],
+    volunteer: [
+      {
+        text: 'NOW',
+        link: 'http://nownyc.org/get-involved/'
+      },
+      {
+        text: 'NARAL',
+        link: 'http://actnow.prochoiceamerica.org/signup/volunteer/?source=website#.WCNXKuErKgQ'
+      },
 
-			{
-				text: "PLANNED PARENTHOOD",
-				link: "https://plannedparenthoodrccorpvolunteer.peoplefluent.com/index.html"
-			},
-			{
-				text: "RAINN",
-				link: "https://www.rainn.org/get-involved"
-			}
-		]
-	},
+      {
+        text: 'PLANNED PARENTHOOD',
+        link: 'https://plannedparenthoodrccorpvolunteer.peoplefluent.com/index.html'
+      },
+      {
+        text: 'RAINN',
+        link: 'https://www.rainn.org/get-involved'
+      }
+    ]
+  },
 
-	earth: {
-		donate: [
-			{
-				text: "EARTHJUSTICE",
-				link: "https://secure.earthjustice.org/site/Donation2?df_id=10960&10960.donation=form1&_ga=1.146521844.1974926048.1478708236"
-			},
-			{
-				text: "THE SIERRA CLUB",
-				link: "https://vault.sierraclub.org/ways-to-give/"
-			},
-			{
-				text: "CLIMATE SCIENCE LEGAL DEFENSE FUND",
-				link: "http://climatesciencedefensefund.org/donate/monthly/"
-			},
-			{
-				text: "LEAGUE OF CONSERVATION VOTERS",
-				link: "https://secure3.convio.net/lcv/site/Donation2?df_id=1741&1741.donation=form1"
-			},
-			{
-				text: "NATURAL RESOURCES DEFENSE COUNCIL",
-				link: "https://secure.nrdconline.org/site/Donation2?df_id=15743&15743.donation=form1&_ga=1.264541580.203263383.1479235887"
-			},
-			{
-				text: "COOL EARTH",
-				link: "https://www.coolearth.org/get-involved/donate/"
-			}
-		],
-		volunteer: [
-			{
-				text: "EARTHJUSTICE",
-				link: "http://earthjustice.org/action"
-			},
-			{
-				text: "NEXTGEN CLIMATE ACTION",
-				link: "https://nextgenclimate.org/volunteer/"
-			},
-			{
-				text: "SAFECAST",
-				link: "http://blog.safecast.org/volunteer/"
-			}
-		]
-	}
+  earth: {
+    donate: [
+      {
+        text: 'EARTHJUSTICE',
+        link: 'https://secure.earthjustice.org/site/Donation2?df_id=10960&10960.donation=form1&_ga=1.146521844.1974926048.1478708236'
+      },
+      {
+        text: 'THE SIERRA CLUB',
+        link: 'https://vault.sierraclub.org/ways-to-give/'
+      },
+      {
+        text: 'CLIMATE SCIENCE LEGAL DEFENSE FUND',
+        link: 'http://climatesciencedefensefund.org/donate/monthly/'
+      },
+      {
+        text: 'LEAGUE OF CONSERVATION VOTERS',
+        link: 'https://secure3.convio.net/lcv/site/Donation2?df_id=1741&1741.donation=form1'
+      },
+      {
+        text: 'NATURAL RESOURCES DEFENSE COUNCIL',
+        link: 'https://secure.nrdconline.org/site/Donation2?df_id=15743&15743.donation=form1&_ga=1.264541580.203263383.1479235887'
+      },
+      {
+        text: 'COOL EARTH',
+        link: 'https://www.coolearth.org/get-involved/donate/'
+      }
+    ],
+    volunteer: [
+      {
+        text: 'EARTHJUSTICE',
+        link: 'http://earthjustice.org/action'
+      },
+      {
+        text: 'NEXTGEN CLIMATE ACTION',
+        link: 'https://nextgenclimate.org/volunteer/'
+      },
+      {
+        text: 'SAFECAST',
+        link: 'http://blog.safecast.org/volunteer/'
+      }
+    ]
+  }
 
 };
