@@ -10,8 +10,8 @@ import './styles/main.scss';
  * Both configureStore and Root are required conditionally.
  * See configureStore.js and Root.js for more details.
  */
-import { configureStore } from './store/configureStore';
-import { Root } from './containers/Root';
+import configureStore from './store/configureStore';
+import Root from './containers/Root';
 
 const store = configureStore();
 
