@@ -2,7 +2,7 @@ import { IDEAS_LIST } from '../constants/IdeasConstants';
 
 export default function ideas(state = IDEAS_LIST, action) {
   switch (action.type) {
-  default:
-    return state;
+    default:
+      return state;
   }
 }
