@@ -1,8 +1,8 @@
-import { LINKS_LIST } from '../constants/LinkConstants';
+import LINKS_LIST from '../constants/LinkConstants';
 
 export default function ideas(state = LINKS_LIST, action) {
   switch (action.type) {
-  default:
-    return state;
+    default:
+      return state;
   }
 }
