@@ -1,7 +1,8 @@
 import { createStore, applyMiddleware, compose } from 'redux';
-import rootReducer from '../reducers';
 import createLogger from 'redux-logger';
 import thunk from 'redux-thunk';
+
+import rootReducer from '../reducers';
 import DevTools from '../containers/DevTools';
 
 /**
@@ -19,15 +20,17 @@ const finalCreateStore = compose(
   DevTools.instrument()
 )(createStore);
 
-module.exports = function configureStore(initialState) {
+function configureStore(initialState) {
   const store = finalCreateStore(rootReducer, initialState);
 
   // Hot reload reducers (requires Webpack or Browserify HMR to be enabled)
   if (module.hot) {
     module.hot.accept('../reducers', () =>
-      store.replaceReducer(require('../reducers'))
+      store.replaceReducer(require('../reducers')) // eslint-disable-line
     );
   }
 
   return store;
-};
+}
+
+export default configureStore;
