@@ -1,11 +1,14 @@
 import { createStore, applyMiddleware, compose } from 'redux';
-import rootReducer from '../reducers';
 import thunk from 'redux-thunk';
+
+import rootReducer from '../reducers';
 
 const finalCreateStore = compose(
   applyMiddleware(thunk)
 )(createStore);
 
-module.exports = function configureStore(initialState) {
+function configureStore(initialState) {
   return finalCreateStore(rootReducer, initialState);
-};
+}
+
+export default configureStore;
