@@ -6,7 +6,7 @@ module.exports = {
   entry: [
     'eventsource-polyfill', // necessary for hot reloading with IE
     'webpack-hot-middleware/client',
-    './src/index'
+    './src/index.jsx'
   ],
   output: {
     path: path.join(__dirname, 'dist'),
@@ -34,17 +34,20 @@ module.exports = {
       'process.env.NODE_ENV': JSON.stringify('development')
     }),
   ],
+  resolve: {
+    extensions: ['.js', '.jsx', '.json']
+  },
   module: {
     loaders: [
       {
-        test: /\.js?/,
+        test: /\.jsx?/,
         exclude: [/node_modules/, /styles/],
-        loaders: ['babel'],
+        loaders: ['babel-loader'],
         include: path.join(__dirname, 'src')
       },
       {
         test: /\.scss$/,
-        loader: 'style!css!sass'
+        loader: 'style-loader!css-loader!sass-loader'
       }
     ]
   }
