@@ -5,7 +5,7 @@ const autoprefixer = require('autoprefixer');
 module.exports = {
   devtool: 'source-map',
   entry: [
-    './src/index'
+    './src/index.jsx'
   ],
   output: {
     path: path.join(__dirname, 'dist', 'static'),
@@ -18,34 +18,34 @@ module.exports = {
      * means is that frequently used IDs will get lower/shorter IDs - so they become
      * more predictable.
      */
-    new webpack.optimize.OccurenceOrderPlugin(),
+    new webpack.optimize.OccurrenceOrderPlugin(),
     /**
      * See description in 'webpack.config.dev' for more info.
      */
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify('production')
-    }),
-    /**
-     * Some of you might recognize this! It minimizes all your JS output of chunks.
-     * Loaders are switched into a minmizing mode. Obviously, you'd only want to run
-     * your production code through this!
-     */
-    new webpack.optimize.UglifyJsPlugin({
-      compressor: {
-        warnings: false
-      }
     })
   ],
+  resolve: {
+    extensions: ['.js', '.jsx', '.json'],
+    alias: {
+      'react': 'preact-compat',
+      'react-dom': 'preact-compat'
+    }
+  },
   module: {
     loaders: [
       {
-        test: /\.js$/,
-        loaders: ['babel'],
-        include: path.join(__dirname, 'src')
+        test: /\.jsx?$/,
+        loaders: ['babel-loader'],
+        include: [
+          path.join(__dirname, 'src'),
+          path.join(__dirname, './node_modules/preact-compat')
+        ]
       },
       {
         test: /\.scss$/,
-        loader: 'style!css!postcss!sass'
+        loader: 'style-loader!css-loader!postcss-loader!sass-loader'
       }
     ]
   }
